+++ lib/src/components/export.dart
@@ -1,8 +1,6 @@
 // APP BAR
 // ignore_for_file: directives_ordering
 
-export 'app_bars/nebux_app_bar.dart';
-
 // BUTTONS
 export 'buttons/export.dart';
 
@@ -22,4 +20,4 @@ export 'scaffold/export.dart';
 export 'shared/export.dart';
 
 // SHIMMERS
-export 'package:app_shimmer/app_shimmer.dart';
+export 'shimmers/export.dart';