+++ lib/src/core/utils/nebux_utils.dart
@@ -1,5 +1,7 @@
 import 'dart:math';
 
+import 'package:flutter/foundation.dart';
+
 class NebuxUtils {
   static String generateRandomTag([String? value]) {
     final random = 100000 + Random().nextInt(900000);
@@ -35,7 +37,7 @@ class NebuxUtils {
   }
 
   /// Checks if the current platform is web
-  static bool get isWeb => identical(0, 0.0);
+  static bool get isWeb => kIsWeb;
 
   /// Checks if the current platform is mobile
   static bool get isMobile => !isWeb;